pyproject.toml
[project]
name = "levindorge"
version = "0.1.2"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.2.0",
    "cleanstack[fastapi,mongo]",
    "fastapi[standard]>=0.117",
    "faststream[cli,redis]==0.5.48",
    "itsdangerous==2.2.0",
    "logfire[fastapi,pymongo]==4.11.0",
    "pillow==11.3.0",
    "pwdlib[argon2,bcrypt]==0.2.1",
    "pydantic-settings==2.11.0",
    "pyjwt==2.10.1",
    "tactill",
    "tenacity==9.1.2",
    "unidecode==1.4.0",
]

[dependency-groups]
dev = [
    "ruff==0.13.3",
    "mypy==1.18.2",
    "pytest==8.4.2",
    "pre-commit==4.3.0",
]

[tool.uv.sources]
tactill = { git = "https://github.com/julienpillaud/tactill.git", tag = "0.6.0" }
cleanstack = { git = "https://github.com/julienpillaud/cleanstack", tag = "0.5.0" }

[tool.ruff]
output-format = "concise"

[tool.ruff.lint]
select = [
    "E", "W",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
    "PL",  # Pylint
    "RUF",  # ruff
]
ignore = [
    "PLR0911",  # Too many return statements
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments in function definition
    "PLR0915",  # Too many statements
]

[tool.mypy]
strict = true
plugins = ['pydantic.mypy']
exclude = [
    '^app/blueprints/',
    '^app/entities/',
    '^app/interfaces/',
    '^app/logging/',
    '^app/repository/',
    '^app/use_cases/',
    '^app/utils/',
    '^app/config.py$',
    '^app/worker.py$',
    '^app/main.py$',

    '^scripts/',
    '^tests/',
]
